Move HTTP handling tests to requests.test.js
