--- github-api/HttpError.js
@@ -1,6 +1,0 @@
-export class HttpError extends Error {
-  constructor ({ status, message }) {
-    super(`${status}: ${message}`)
-    this.name = this.constructor.name
-  }
-}
