+++ github-api/requests.js
@@ -1,5 +1,4 @@
 import fetch from 'cross-fetch'
-import { HttpError } from './HttpError'
 
 const url = 'https://api.github.com'
 
@@ -22,6 +21,13 @@ export async function post (path, accessToken, body) {
   })
   await handleError(response)
   return response.json()
+}
+
+export class HttpError extends Error {
+  constructor ({ status, message }) {
+    super(`${status}: ${message}`)
+    this.name = this.constructor.name
+  }
 }
 
 function headers (accessToken) {
